+++ src/BaseballGame.js
@@ -1,8 +1,7 @@
 import showResult from "./showResult.js";
-import init from "./init.js";
 
 export default function BaseballGame() {
-    function play(computerInputNumbers, userInputNumbers) {
+    this.play = function (computerInputNumbers, userInputNumbers) {
         let strike = 0;
         let ball = 0;
         for (let i = 0; i < userInputNumbers.length; i++) {
@@ -16,6 +15,5 @@ export default function BaseballGame() {
         }
         const result = showResult(ball, strike);
         return result;
-    }
-    init(play);
+    };
 }
